Rename validators.ts to types.ts
